requirements.txt
pillow
opencv-python-headless
reedsolo
numpy